split out cli from core features
